Provide Go language scripts to create NetBackup policies in one step and with step by step
