+++ README.md
@@ -31,10 +31,11 @@ Use the following commands to run the curl samples.
 - `./get_nb_jobs.sh -master <master_server> -username <username> -password <password>`
 - `./get_nb_images.sh -master <master_server> -username <username> -password <password>`
 
-#### Executing the snippets using go
+#### Executing the recipes using go (often referred to as golang)
 Pre-requisites:
 - NetBackup 8.1.2 or higher
 - go1.10.2 or higher
 
 Use the following commands to run the go samples.
 - `go run ./create_policy_step_by_step.go -nbmaster <master_server> -username <username> -password <password>`
+- `go run ./create_policy_in_one_step.go -nbmaster <master_server> -username <username> -password <password>`
